Updated shipcontrol

+++ inc/shipcontrol.h
@@ -22,7 +22,7 @@ typedef struct {
 
 //------------------------------Functions---------------------------
 void shipSetup(ship_t * ship);
-int shipControl(ship_t * ship);
+uint8_t shipControl(ship_t * ship);
 void printShip(ship_t * ship);
 void bulletSetup(bullet_t bullet[]);
 void updateBullet(ship_t * ship, bullet_t * bullet, uint8_t flagbullet, uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2);
